4 and mini version of 5
Don't pay attention to the name first_30_rows, that variable contains the first 50 rows of the df dataframe. DON'T BUILD THE ALL_BIGRAMS_BNC, IT WILL TAKE FOREVER (actually I cannot push this file because it has almost 1 GB of information so u may have to create it) It is already in a json file that can be uploaded by the code given in task 4. Be aware, that the process of loading the data will take some time and it may crash your computer if u have too many tabs and windows open. I can explain this better tomorrow, but now I am exhausted.